Restructure taxonomy extraction
- Replace out-parameter tree-walk recursion with generators yielding
  typed rows (DefinitionRow/PresentationRow); callers now filter and
  serialise explicitly instead of via includeUsable flags and
  variable-shape tuples
- Split taxonomy_info.py into a thin Arelle plugin shim and a new
  taxonomy_extraction.py; extractors return their data instead of
  writing into plugin data, making them unit-testable without the
  plugin machinery
- Rename convert_recursive -> convertRecursive
